Add generic TT plugin execution hooks Move TT execution, DP orchestration, launcher behavior, and config access behind generic plugin-selected classes so shared vLLM core no longer owns TT-specific engine paths.
