+++ plugins/vllm-tt-plugin/src/vllm_tt_plugin/engine.py
@@ -1,0 +1,624 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from __future__ import annotations
+
+import pickle
+from concurrent.futures import Future
+from dataclasses import dataclass
+from typing import Any, TypeVar, cast
+
+import torch
+import torch.distributed as dist
+
+from vllm.logger import init_logger
+from vllm.v1.core.sched.output import SchedulerOutput
+from vllm.v1.engine import EngineCoreOutputs
+from vllm.v1.engine.core import DPEngineCoreProc, EngineCore, EngineCoreProc
+from vllm.v1.outputs import EMPTY_MODEL_RUNNER_OUTPUT, ModelRunnerOutput
+from vllm_tt_plugin.scheduler import TTSchedulingMode
+
+logger = init_logger(__name__)
+_T = TypeVar("_T")
+
+
+@dataclass
+class DPGatherHandle:
+    future: Future[tuple[torch.Tensor, list]]
+    scheduler_output: SchedulerOutput | None
+    local_has_requests: bool
+    is_decode: bool
+    overlap_ok: bool
+    any_needs_logprobs: bool
+    req_ids: list[str]
+    req_id_to_index: dict[str, int]
+
+
+class TTExecutionMixin:
+    """TT non-DP execution policy for plugin-selected engine cores."""
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)  # type: ignore[misc]
+        if (
+            self.batch_queue is not None
+            and self.vllm_config.scheduler_config.async_scheduling
+            and self.vllm_config.parallel_config.data_parallel_size == 1
+        ):
+            self.step_fn = self.step_with_batch_queue_tt
+
+    def step(self) -> tuple[dict[int, EngineCoreOutputs], bool]:
+        """TT regular execution path.
+
+        TT's scheduler computes structured-output grammar state before
+        execution, and TT sampling runs inside ``execute_model``.
+        """
+        if self._scheduler_paused:
+            return {}, False
+
+        if not self.scheduler.has_requests():
+            return {}, False
+
+        scheduler_output = self.scheduler.schedule()
+        future = self.model_executor.execute_model(scheduler_output, non_block=True)
+        with (
+            self.log_error_detail(scheduler_output),
+            self.log_iteration_details(scheduler_output),
+        ):
+            model_output = future.result()
+            if model_output is None:
+                model_output = self.model_executor.sample_tokens(None)
+
+        self._process_aborts_queue()
+        engine_core_outputs = self.scheduler.update_from_output(
+            scheduler_output, model_output
+        )
+        return engine_core_outputs, scheduler_output.total_num_scheduled_tokens > 0
+
+    def step_with_batch_queue_tt(
+        self,
+    ) -> tuple[dict[int, EngineCoreOutputs] | None, bool]:
+        """TT-specific async batch-queue path for non-DP execution."""
+        batch_queue = self.batch_queue
+        assert batch_queue is not None
+
+        # Match the shared queue behavior: prefer keeping the queue filled before
+        # blocking on the oldest in-flight result.
+        assert len(batch_queue) < self.batch_queue_size
+
+        model_executed = False
+        deferred_scheduler_output: SchedulerOutput | None = None
+        if self.scheduler.has_requests():
+            scheduler_output = self.scheduler.schedule()
+            if not self.is_ec_producer:
+                model_executed = scheduler_output.total_num_scheduled_tokens > 0
+
+            if self.is_pooling_model or not model_executed:
+                future = cast(
+                    Future[ModelRunnerOutput],
+                    self.model_executor.execute_model(
+                        scheduler_output, non_block=True
+                    ),
+                )
+            elif scheduler_output.pending_structured_output_tokens:
+                # TT consumes structured-output state inside execute_model(), so
+                # the grammar bitmask must be populated before submission.
+                deferred_scheduler_output = scheduler_output
+            else:
+                future = cast(
+                    Future[ModelRunnerOutput],
+                    self.model_executor.execute_model(
+                        scheduler_output, non_block=True
+                    ),
+                )
+
+            if deferred_scheduler_output is None:
+                batch_queue.appendleft((future, scheduler_output, future))
+                if (
+                    model_executed
+                    and len(batch_queue) < self.batch_queue_size
+                    and not batch_queue[-1][0].done()
+                ):
+                    return None, True
+
+        elif not batch_queue:
+            return None, False
+
+        future, scheduler_output, _ = batch_queue.pop()
+        with self.log_error_detail(scheduler_output):
+            model_output = future.result()
+
+        engine_core_outputs = self.scheduler.update_from_output(
+            scheduler_output, model_output
+        )
+
+        if deferred_scheduler_output is not None:
+            self.scheduler.get_grammar_bitmask(deferred_scheduler_output)
+            future = cast(
+                Future[ModelRunnerOutput],
+                self.model_executor.execute_model(
+                    deferred_scheduler_output, non_block=True
+                ),
+            )
+            batch_queue.appendleft((future, deferred_scheduler_output, future))
+
+        return engine_core_outputs, model_executed
+
+
+class TTEngineCore(TTExecutionMixin, EngineCore):
+    """In-process TT engine core."""
+
+
+class TTEngineCoreProc(TTExecutionMixin, EngineCoreProc):
+    """Multiprocessing TT engine core."""
+
+
+class TTDPEngineCoreProc(DPEngineCoreProc):
+    """TT data-parallel engine core with gathered-batch orchestration."""
+
+    def _dp_any_rank_has_scheduler_requests(self) -> bool:
+        local_has_requests = 1 if self.scheduler.has_requests() else 0
+        has_requests_t = torch.tensor([local_has_requests], dtype=torch.int32)
+        try:
+            dist.all_reduce(has_requests_t, op=dist.ReduceOp.SUM, group=self.dp_group)
+        except RuntimeError as e:
+            # During shutdown, peers may close connections mid-collective.
+            if "Connection closed by peer" in str(e):
+                logger.debug("Collective failed during shutdown, exiting gracefully")
+                raise SystemExit() from e
+            raise
+        return int(has_requests_t.item()) > 0
+
+    def _dp_negotiate_forced_mode(self) -> TTSchedulingMode:
+        has_running = bool(getattr(self.scheduler, "running", []))
+        has_waiting = bool(getattr(self.scheduler, "waiting", False))
+        max_running = getattr(self.scheduler, "max_num_running_reqs", 0)
+        has_capacity = len(getattr(self.scheduler, "running", [])) < max_running
+        local_prefill_intent = (
+            1 if (has_waiting and ((not has_running) or has_capacity)) else 0
+        )
+        intent_tensor = torch.tensor([local_prefill_intent], dtype=torch.int32)
+        self.dlog("before_intent_allreduce intent_tensor=%s", intent_tensor)
+        dist.all_reduce(intent_tensor, op=dist.ReduceOp.MAX, group=self.dp_group)
+        forced_mode = TTSchedulingMode.from_prefill_intent(int(intent_tensor.item()))
+        self.dlog("after_intent_allreduce forced_mode=%s", forced_mode)
+        self._dp_gather_forced_mode = forced_mode
+        return forced_mode
+
+    def _dp_apply_forced_mode(self, forced_mode: TTSchedulingMode) -> None:
+        set_mode = getattr(self.scheduler, "set_forced_mode", None)
+        if callable(set_mode):
+            set_mode(forced_mode)
+
+    def step(self) -> tuple[dict[int, EngineCoreOutputs], bool]:
+        if self._scheduler_paused:
+            return {}, False
+
+        if not self._dp_any_rank_has_scheduler_requests():
+            return {}, False
+
+        forced_mode = self._dp_negotiate_forced_mode()
+        if not self.scheduler.has_requests():
+            _ = self._execute_model_dp_gather(None)
+            return {}, False
+
+        self._dp_apply_forced_mode(forced_mode)
+        scheduler_output = self.scheduler.schedule()
+        self._dp_apply_forced_mode(TTSchedulingMode.DEFAULT)
+
+        model_output = self._execute_model_dp_gather(scheduler_output)
+        self._process_aborts_queue()
+        engine_core_outputs = self.scheduler.update_from_output(
+            scheduler_output, model_output
+        )
+        return engine_core_outputs, scheduler_output.total_num_scheduled_tokens > 0
+
+    def step_dp_with_batch_queue(
+        self,
+    ) -> tuple[dict[int, EngineCoreOutputs] | None, bool]:
+        assert self.batch_queue is not None
+
+        global_has_requests = self._dp_any_rank_has_scheduler_requests()
+        prev_handle = self._dp_in_flight
+        if not global_has_requests and prev_handle is None:
+            return {}, False
+
+        forced_mode = TTSchedulingMode.DEFAULT
+        scheduler_output: SchedulerOutput | None = None
+        model_executed = False
+        current_overlap_ok = False
+        if global_has_requests:
+            forced_mode = self._dp_negotiate_forced_mode()
+            if self.scheduler.has_requests():
+                self._dp_apply_forced_mode(forced_mode)
+                scheduler_output = self.scheduler.schedule()
+                self._dp_apply_forced_mode(TTSchedulingMode.DEFAULT)
+                if not self.is_ec_producer:
+                    model_executed = (
+                        scheduler_output.total_num_scheduled_tokens > 0
+                    )
+            if forced_mode == TTSchedulingMode.DECODE_ONLY:
+                current_overlap_ok = self._dp_can_attempt_steady_decode_from_scheduler(
+                    scheduler_output
+                )
+
+        def _finalize_previous(
+            handle: DPGatherHandle,
+        ) -> dict[int, EngineCoreOutputs]:
+            model_output = self.dp_gather_finalize(handle)
+            if handle.scheduler_output is None:
+                return {}
+            return self.scheduler.update_from_output(
+                handle.scheduler_output, model_output
+            )
+
+        finalize_before_submit = prev_handle is not None and (
+            not global_has_requests
+            or not prev_handle.overlap_ok
+            or not current_overlap_ok
+        )
+
+        engine_core_outputs: dict[int, EngineCoreOutputs] | None = {}
+        if finalize_before_submit:
+            assert prev_handle is not None
+            engine_core_outputs = _finalize_previous(prev_handle)
+            prev_handle = None
+
+        if (
+            scheduler_output is not None
+            and scheduler_output.pending_structured_output_tokens
+        ):
+            self.scheduler.get_grammar_bitmask(scheduler_output)
+
+        next_handle: DPGatherHandle | None = None
+        if global_has_requests:
+            next_handle = self.dp_gather_submit(
+                scheduler_output,
+                overlap_ok=current_overlap_ok,
+            )
+
+        if not finalize_before_submit and prev_handle is not None:
+            engine_core_outputs = _finalize_previous(prev_handle)
+
+        self._dp_in_flight = next_handle
+
+        if not global_has_requests:
+            return engine_core_outputs, False
+
+        return engine_core_outputs, model_executed
+
+    def _dp_can_attempt_steady_decode_from_scheduler(
+        self,
+        scheduler_output: SchedulerOutput | None,
+    ) -> bool:
+        local_overlap_ok = int(
+            self.model_executor.collective_rpc(
+                "can_attempt_steady_dp_decode_from_scheduler",
+                args=(scheduler_output,),
+            )[0]
+        )
+        overlap_ok_t = torch.tensor([local_overlap_ok], dtype=torch.int32)
+        dist.all_reduce(overlap_ok_t, op=dist.ReduceOp.MIN, group=self.dp_group)
+        overlap_ok = bool(overlap_ok_t.item())
+        self.dlog("steady_decode_overlap_ok=%s", overlap_ok)
+        return overlap_ok
+
+    def dp_gather_submit(
+        self,
+        scheduler_output: SchedulerOutput | None,
+        *,
+        overlap_ok: bool = False,
+    ) -> DPGatherHandle:
+        parallel_config = self.vllm_config.parallel_config
+        group = self.dp_group
+        rank = self.dp_rank
+        local_rank = parallel_config.data_parallel_rank_local
+        world = parallel_config.data_parallel_size
+
+        local_has_requests = scheduler_output is not None
+        if scheduler_output is not None:
+            self.dlog(
+                "enter_gather tokens=%d",
+                scheduler_output.total_num_scheduled_tokens,
+            )
+
+        assert hasattr(self, "_dp_gather_forced_mode"), "forced_mode not set"
+        is_decode = self._dp_gather_forced_mode == TTSchedulingMode.DECODE_ONLY
+
+        all_local_inputs = self.model_executor.collective_rpc(
+            "build_dp_model_input", args=(scheduler_output,)
+        )[0]
+        (
+            local_input,
+            local_max_blocks,
+            local_has_structured,
+            local_has_penalties,
+            local_reset_batch,
+            local_can_sample_device,
+            local_needs_logprobs,
+            req_ids,
+            req_id_to_index,
+        ) = all_local_inputs
+        max_blocks_decode = None
+        any_structured_inputs = False
+        any_needs_logprobs = False
+
+        gathered_inputs: Any = None
+        if is_decode:
+            input_info_t = torch.tensor(
+                [
+                    local_max_blocks,
+                    local_has_structured,
+                    local_has_penalties,
+                    local_reset_batch,
+                    1 - local_can_sample_device,
+                    local_needs_logprobs,
+                ],
+                dtype=torch.int32,
+            )
+            dist.all_reduce(input_info_t, op=dist.ReduceOp.MAX, group=group)
+            max_blocks_decode = int(input_info_t[0].item())
+            any_structured_inputs = input_info_t[1].item() > 0
+            any_penalties_inputs = input_info_t[2].item() > 0
+            any_reset_batch = input_info_t[3].item() > 0
+            all_sample_device = input_info_t[4].item() == 0
+            any_needs_logprobs = input_info_t[5].item() > 0
+
+            decode_inputs: dict[str, Any] = self.model_executor.collective_rpc(
+                "build_dp_decode_gather_input",
+                args=(
+                    local_input,
+                    max_blocks_decode,
+                    any_structured_inputs,
+                    any_penalties_inputs,
+                ),
+            )[0]
+
+            int_local = decode_inputs["int_inputs"]
+            float_local = decode_inputs["float_inputs"]
+
+            stacked_int = None
+            stacked_float = None
+            gather_list_int = None
+            gather_list_float = None
+            if rank == 0:
+                stacked_int = torch.empty(
+                    (world, *int_local.shape), dtype=int_local.dtype
+                )
+                stacked_float = torch.empty(
+                    (world, *float_local.shape), dtype=float_local.dtype
+                )
+                gather_list_int = [stacked_int[i] for i in range(world)]
+                gather_list_float = [stacked_float[i] for i in range(world)]
+
+            dist.gather(int_local, gather_list_int, dst=0, group=group)
+            dist.gather(float_local, gather_list_float, dst=0, group=group)
+            if len(self.dp_device_ranks) > 1:
+                if rank == 0:
+                    for dst in self.dp_device_ranks[1:]:
+                        dist.send(stacked_int, dst=dst, group=group)
+                        dist.send(stacked_float, dst=dst, group=group)
+                elif local_rank == 0:
+                    stacked_int = torch.empty(
+                        (world, *int_local.shape), dtype=int_local.dtype
+                    )
+                    stacked_float = torch.empty(
+                        (world, *float_local.shape), dtype=float_local.dtype
+                    )
+                    dist.recv(stacked_int, src=0, group=group)
+                    dist.recv(stacked_float, src=0, group=group)
+
+            gathered_tokens_inputs = None
+            if any_penalties_inputs and (not all_sample_device or any_reset_batch):
+                if rank == 0:
+                    gathered_tokens_inputs = [None for _ in range(world)]
+                local_tokens_inputs = decode_inputs["sampling_tokens_inputs"]
+                dist.gather_object(
+                    local_tokens_inputs, gathered_tokens_inputs, dst=0, group=group
+                )
+
+                if len(self.dp_device_ranks) > 1:
+                    if rank == 0:
+                        pickled_tokens = pickle.dumps(gathered_tokens_inputs)
+                        tokens_tensor = torch.frombuffer(
+                            pickled_tokens, dtype=torch.uint8
+                        )
+                        tokens_size = torch.tensor(
+                            [tokens_tensor.numel()], dtype=torch.long
+                        )
+                        for dst in self.dp_device_ranks[1:]:
+                            dist.send(tokens_size, dst=dst, group=group)
+                            dist.send(tokens_tensor, dst=dst, group=group)
+                    elif local_rank == 0:
+                        tokens_size = torch.zeros(1, dtype=torch.long)
+                        dist.recv(tokens_size, src=0, group=group)
+                        tokens_tensor = torch.empty(
+                            tokens_size.item(), dtype=torch.uint8
+                        )
+                        dist.recv(tokens_tensor, src=0, group=group)
+                        gathered_tokens_inputs = pickle.loads(
+                            tokens_tensor.numpy().tobytes()
+                        )
+
+            gathered_host_only_sample_params = None
+            if not all_sample_device:
+                if rank == 0:
+                    gathered_host_only_sample_params = [None for _ in range(world)]
+                local_host_only_sample_params = decode_inputs.get(
+                    "host_only_sample_params"
+                )
+                dist.gather_object(
+                    local_host_only_sample_params,
+                    gathered_host_only_sample_params,
+                    dst=0,
+                    group=group,
+                )
+
+                if len(self.dp_device_ranks) > 1:
+                    if rank == 0:
+                        pickled_host_only = pickle.dumps(
+                            gathered_host_only_sample_params
+                        )
+                        host_only_tensor = torch.frombuffer(
+                            pickled_host_only, dtype=torch.uint8
+                        )
+                        host_only_size = torch.tensor(
+                            [host_only_tensor.numel()], dtype=torch.long
+                        )
+                        for dst in self.dp_device_ranks[1:]:
+                            dist.send(host_only_size, dst=dst, group=group)
+                            dist.send(host_only_tensor, dst=dst, group=group)
+                    elif local_rank == 0:
+                        host_only_size = torch.zeros(1, dtype=torch.long)
+                        dist.recv(host_only_size, src=0, group=group)
+                        host_only_tensor = torch.empty(
+                            host_only_size.item(), dtype=torch.uint8
+                        )
+                        dist.recv(host_only_tensor, src=0, group=group)
+                        gathered_host_only_sample_params = pickle.loads(
+                            host_only_tensor.numpy().tobytes()
+                        )
+
+            if local_rank == 0:
+                gathered_inputs = {
+                    "int_inputs": stacked_int,
+                    "float_inputs": stacked_float,
+                    "sampling_tokens_inputs": gathered_tokens_inputs,
+                    "host_only_sample_params": gathered_host_only_sample_params,
+                    "reset_batch": any_reset_batch,
+                    "all_sample_device": all_sample_device,
+                }
+
+        else:
+            gathered_inputs = None
+            if rank == 0:
+                gathered_inputs = [None for _ in range(world)]
+
+            logprobs_flag_t = torch.tensor([local_needs_logprobs], dtype=torch.int32)
+            dist.all_reduce(logprobs_flag_t, op=dist.ReduceOp.MAX, group=group)
+            any_needs_logprobs = logprobs_flag_t[0].item() > 0
+
+            dist.gather_object(local_input, gathered_inputs, dst=0, group=group)
+            if len(self.dp_device_ranks) > 1:
+                if rank == 0:
+                    pickled_data = pickle.dumps(gathered_inputs)
+                    object_tensor = torch.frombuffer(pickled_data, dtype=torch.uint8)
+                    size_tensor = torch.tensor(
+                        [object_tensor.numel()], dtype=torch.long
+                    )
+                    for dst in self.dp_device_ranks[1:]:
+                        dist.send(size_tensor, dst=dst, group=group)
+                        dist.send(object_tensor, dst=dst, group=group)
+                elif local_rank == 0:
+                    size_tensor = torch.zeros(1, dtype=torch.long)
+                    dist.recv(size_tensor, src=0, group=group)
+                    object_tensor = torch.empty(
+                        size_tensor.item(), dtype=torch.uint8
+                    )
+                    dist.recv(object_tensor, src=0, group=group)
+                    gathered_inputs = pickle.loads(object_tensor.numpy().tobytes())
+        self.dlog("after_inputs_gather")
+
+        should_submit = is_decode or (
+            isinstance(gathered_inputs, list)
+            and any(x is not None for x in gathered_inputs)
+        )
+        if should_submit:
+            collective_future = cast(
+                Future[list[tuple[torch.Tensor, list]]],
+                self.model_executor.collective_rpc(
+                    "concat_and_execute_dp",
+                    args=(
+                        gathered_inputs,
+                        is_decode,
+                        max_blocks_decode,
+                        any_structured_inputs,
+                    ),
+                    kwargs={"non_block": True},
+                    non_block=True,
+                ),
+            )
+            future = _unwrap_single_worker_future(collective_future)
+        else:
+            future = self._completed_dp_gather_future()
+
+        return DPGatherHandle(
+            future=future,
+            scheduler_output=scheduler_output,
+            local_has_requests=local_has_requests,
+            is_decode=is_decode,
+            overlap_ok=overlap_ok,
+            any_needs_logprobs=any_needs_logprobs,
+            req_ids=req_ids,
+            req_id_to_index=req_id_to_index,
+        )
+
+    def dp_gather_finalize(self, handle: DPGatherHandle) -> ModelRunnerOutput:
+        parallel_config = self.vllm_config.parallel_config
+        group = self.dp_group
+        rank = self.dp_rank
+        world = parallel_config.data_parallel_size
+        logprobs_per_dp: list = [None] * world
+
+        result = handle.future.result()
+        assert isinstance(result, tuple) and len(result) == 2
+        send_tensor, logprobs_per_dp = result
+        assert isinstance(send_tensor, torch.Tensor)
+
+        my_ids = torch.empty_like(send_tensor[0])
+        scatter_list = None
+        if rank == 0:
+            scatter_list = [send_tensor[i] for i in range(world)]
+        dist.scatter(my_ids, scatter_list, src=0, group=group)
+        self.dlog("after_results_gather my_ids_shape=%s", tuple(my_ids.shape))
+
+        my_logprobs_val = None
+        if handle.any_needs_logprobs:
+            my_logprobs: list = [None]
+            logprobs_scatter_list = logprobs_per_dp if rank == 0 else None
+            dist.scatter_object_list(
+                my_logprobs, logprobs_scatter_list, src=0, group=group
+            )
+            my_logprobs_val = my_logprobs[0]
+
+        if handle.local_has_requests:
+            output: ModelRunnerOutput = self.model_executor.collective_rpc(
+                "apply_dp_execution_result",
+                args=(
+                    my_ids,
+                    my_logprobs_val,
+                    handle.req_ids,
+                    handle.req_id_to_index,
+                ),
+            )[0]
+            return output
+        return EMPTY_MODEL_RUNNER_OUTPUT
+
+    def _execute_model_dp_gather(
+        self, scheduler_output: SchedulerOutput | None
+    ) -> ModelRunnerOutput:
+        handle = self.dp_gather_submit(scheduler_output, overlap_ok=False)
+        return self.dp_gather_finalize(handle)
+
+    def _completed_dp_gather_future(self) -> Future[tuple[torch.Tensor, list]]:
+        parallel_config = self.vllm_config.parallel_config
+        world = parallel_config.data_parallel_size
+        batch_size = self.vllm_config.scheduler_config.max_num_seqs
+        future: Future[tuple[torch.Tensor, list]] = Future()
+        future.set_result(
+            (torch.zeros((world, batch_size, 1), dtype=torch.int32), [None] * world)
+        )
+        return future
+
+
+def _unwrap_single_worker_future(future: Future[list[_T]]) -> Future[_T]:
+    single_future: Future[_T] = Future()
+
+    def _set_single_result(done_future: Future[list[_T]]) -> None:
+        try:
+            results = done_future.result()
+            assert len(results) == 1
+            single_future.set_result(results[0])
+        except Exception as exc:
+            single_future.set_exception(exc)
+
+    future.add_done_callback(_set_single_result)
+    return single_future
