+++ plugins/vllm-tt-plugin/src/vllm_tt_plugin/launcher.py
@@ -15,16 +15,95 @@ import yaml
 
 from vllm.config import VllmConfig
 from vllm.logger import init_logger
+from vllm.utils.import_utils import resolve_obj_by_qualname
 from vllm.utils.network_utils import get_ip
 from vllm.utils.system_utils import kill_process_tree
-from vllm.v1.engine.core import EngineCoreProc
+from vllm.v1.engine.utils import CoreEngine, CoreEngineLauncher, EngineLaunchPlan
 from vllm.v1.executor.abstract import UniProcExecutor
+from vllm_tt_plugin.config import get_tt_config
 
 logger = init_logger(__name__)
 
 
+class TTLaunchPlan(EngineLaunchPlan):
+    rank_binding_file: str | None = None
+
+
+class TTCoreEngineLauncher(CoreEngineLauncher):
+    def prepare_launch(self, vllm_config: VllmConfig) -> EngineLaunchPlan:
+        rank_binding_file, non_device_dp_ranks = parse_tt_mpi_params(vllm_config)
+        if rank_binding_file is None:
+            return EngineLaunchPlan()
+
+        vllm_config.parallel_config.data_parallel_size_local = len(
+            non_device_dp_ranks
+        )
+        plan = TTLaunchPlan(
+            remote_launched=True,
+            non_device_dp_ranks=non_device_dp_ranks,
+        )
+        plan.rank_binding_file = rank_binding_file
+        return plan
+
+    def get_engines_to_handshake(
+        self,
+        vllm_config: VllmConfig,
+        local_engine_count: int,
+        local_start_index: int | None,
+        dp_rank: int,
+        dp_size: int,
+        local_engines_only: bool,
+        offline_mode: bool,
+        plan: EngineLaunchPlan,
+    ) -> list[CoreEngine]:
+        if not plan.remote_launched:
+            return super().get_engines_to_handshake(
+                vllm_config,
+                local_engine_count,
+                local_start_index,
+                dp_rank,
+                dp_size,
+                local_engines_only,
+                offline_mode,
+                plan,
+            )
+        non_device_dp_ranks = plan.non_device_dp_ranks or set()
+        return [
+            CoreEngine(index=i, local=(i in non_device_dp_ranks))
+            for i in range(vllm_config.parallel_config.data_parallel_size)
+        ]
+
+    def launch_remote_engines(
+        self,
+        plan: EngineLaunchPlan,
+        handshake_address: str,
+        vllm_config: VllmConfig,
+        log_stats: bool,
+        cleanup_target: object,
+    ) -> None:
+        assert isinstance(plan, TTLaunchPlan)
+        assert plan.rank_binding_file is not None
+        assert vllm_config.parallel_config.data_parallel_rank == 0, (
+            "TT MPI must be launched from rank 0"
+        )
+        logger.info(
+            "TT-MPI mixed launch: dp_size=%d local_count=%d "
+            "non_device_ranks=%s handshake=%s",
+            vllm_config.parallel_config.data_parallel_size,
+            vllm_config.parallel_config.data_parallel_size_local,
+            plan.non_device_dp_ranks,
+            handshake_address,
+        )
+        tt_run_launch(
+            handshake_address=handshake_address,
+            vllm_config=vllm_config,
+            rank_binding_file=plan.rank_binding_file,
+            log_stats=log_stats,
+            cleanup_target=cleanup_target,
+        )
+
+
 def _validate_launch_from_rank0_host(mpi_args: str, host_ip: str) -> None:
-    # Parse --map-by to locate rankfile and ensure rank 0 host matches host_ip.
     if not mpi_args:
         return
     try:
@@ -34,17 +113,14 @@ def _validate_launch_from_rank0_host(mpi_args: str, host_ip: str) -> None:
     mapby_path = None
     for i, tok in enumerate(argv):
         if tok.startswith("--map-by"):
-            # Support either --map-by=<VALUE> or "--map-by <VALUE>"
             if "=" in tok:
                 value = tok.split("=", 1)[1]
             elif i + 1 < len(argv):
                 value = argv[i + 1]
             else:
                 value = ""
-            # Extract file=... from VALUE like "rankfile:file=/path"
             if "file=" in value:
                 mapby_path = value.split("file=", 1)[1].split(",", 1)[0].strip()
-            # If value itself is a path, accept it as fallback
             if not mapby_path and value and os.path.isfile(value):
                 mapby_path = value
             break
@@ -55,19 +131,17 @@ def _validate_launch_from_rank0_host(mpi_args: str, host_ip: str) -> None:
     try:
         with open(mapby_path) as f:
             for line in f:
-                s = line.strip()
-                if not s or s.startswith("#"):
+                stripped = line.strip()
+                if not stripped or stripped.startswith("#"):
                     continue
-                # Format: rank 0=HOST slot=...
-                tokens = s.split()
+                tokens = stripped.split()
                 if len(tokens) >= 2 and tokens[0] == "rank" and "=" in tokens[1]:
-                    # left: "0", right: "HOST"
                     left, right = tokens[1].split("=", 1)
                     try:
-                        rnum = int(left)
+                        rank_num = int(left)
                     except Exception:
-                        rnum = -1
-                    if rnum == 0:
+                        rank_num = -1
+                    if rank_num == 0:
                         rank0_host = right.strip()
                         break
     except Exception:
@@ -76,10 +150,8 @@ def _validate_launch_from_rank0_host(mpi_args: str, host_ip: str) -> None:
         return
 
     resolved_ips: set[str] = set()
-    # If rank0_host is already an IP address, add it directly
     if all(c.isdigit() or c == "." for c in rank0_host):
         resolved_ips.add(rank0_host)
-    # Also try to resolve hostname to IP addresses
     try:
         info = socket.getaddrinfo(rank0_host, None, proto=socket.IPPROTO_TCP)
         for ai in info:
@@ -88,33 +160,23 @@ def _validate_launch_from_rank0_host(mpi_args: str, host_ip: str) -> None:
         logger.warning(
             "Failed to resolve IP address for rank 0 host %s: %s", rank0_host, e
         )
-        # If resolution failed and rank0_host is not an IP, we can't validate
         if not resolved_ips:
             return
     assert host_ip in resolved_ips, (
         f"MPI rank 0 host {rank0_host} from rankfile {mapby_path} "
         f"(resolves to {sorted(resolved_ips)}) does not match "
-        f"launcher host IP {host_ip} (must launch from rank 0 host "
-        f"{rank0_host})."
+        f"launcher host IP {host_ip}."
     )
     logger.info("Validated launching from MPI rank 0 host %s", rank0_host)
 
 
 def parse_tt_mpi_params(vllm_config: VllmConfig) -> tuple[str | None, set[int]]:
-    """
-    Parse override_tt_config for a rank binding file (required for launching
-    TT MPI processes), and compute device and local DP ranks.
-    Returns tuple with:
-      - rank_binding_file: str
-      - non_device_dp_ranks: set[int]
-    """
-
     parallel_config = vllm_config.parallel_config
     assert parallel_config.data_parallel_backend != "ray", (
         "TT does not support ray-based data parallel backend"
     )
     dp_size = parallel_config.data_parallel_size
-    override_tt_config = vllm_config.model_config.override_tt_config or {}
+    override_tt_config = get_tt_config(vllm_config)
     rank_binding_file = override_tt_config.get("rank_binding")
     non_device_dp_ranks: set[int] = set()
     if rank_binding_file:
@@ -135,12 +197,9 @@ def parse_tt_mpi_params(vllm_config: VllmConfig) -> tuple[str | None, set[int]]:
                 f"data_parallel_size ({dp_size}) must be divisible by number "
                 f"of device MPI ranks ({mpi_world})"
             )
-        # Assume DP world is evenly split into mpi_world groups and set
-        # device DP ranks as the first rank in each group.
         dp_size_per_mpi_rank = dp_size // mpi_world
         device_dp_ranks = {i * dp_size_per_mpi_rank for i in range(mpi_world)}
         non_device_dp_ranks = {i for i in range(dp_size) if i not in device_dp_ranks}
-
     return rank_binding_file, non_device_dp_ranks
 
 
@@ -150,29 +209,8 @@ def tt_run_launch(
     rank_binding_file: str,
     log_stats: bool,
     cleanup_target: object,
-):
-    """
-    Launch TT MPI processes via tt-run from rank 0.
-    Uses args from override_tt_config:
-      - rank_binding: str (required, already parsed)
-      - mpi_args: str (optional, parsed here)
-      - extra_ttrun_args: str (optional, parsed here)
-      - config_pkl_dir: str (required, parsed here)
-    Args:
-        handshake_address: ZMQ address for engine handshake communication.
-        vllm_config: Configuration object containing model and parallel config.
-        rank_binding_file: Path to YAML file specifying MPI rank bindings.
-        log_stats: Whether to enable statistics logging in engine processes.
-        cleanup_target: Object whose lifecycle determines when to clean up
-            the MPI subprocess. A finalizer will be set up automatically.
-    """
-
-    assert rank_binding_file and isinstance(rank_binding_file, str), (
-        "rank_binding must be provided to tt_run_launch as a non-empty string"
-    )
-
-    # Parse override_tt_config for optional fields.
-    override_tt_config = vllm_config.model_config.override_tt_config or {}
+) -> None:
+    override_tt_config = get_tt_config(vllm_config)
     mpi_args = override_tt_config.get("mpi_args", "")
     extra_ttrun_args = override_tt_config.get("extra_ttrun_args")
     cfg_dir = override_tt_config.get("config_pkl_dir")
@@ -185,29 +223,20 @@ def tt_run_launch(
         raise RuntimeError("override_tt_config['config_pkl_dir'] must be a directory")
 
     host_ip = get_ip()
-    # Data parallel master IP must be the same as the launcher host's IP since
-    # the launcher binds the rendezvous port and sends it to all engines.
     parallel_config = vllm_config.parallel_config
     assert parallel_config.data_parallel_master_ip == host_ip, (
         f"data_parallel_master_ip {parallel_config.data_parallel_master_ip} "
-        f"must be the same as the launcher host's IP {host_ip}"
+        f"must be the same as launcher host IP {host_ip}"
     )
-    # Launch must be done on host with MPI rank 0 since we are setting that
-    # process's DP rank to 0 and torch distributed uses DP rank 0 to bind TCP
-    # rendezvous endpoint.
     _validate_launch_from_rank0_host(mpi_args, host_ip)
 
-    # Serialize vllm_config for remote engines to load.
     serialized_config_path = os.path.join(cfg_dir, "tmp_vllm_tt_cfg.pkl")
     with open(serialized_config_path, "wb") as tf:
         cloudpickle.dump(vllm_config, tf)
 
-    # Create a temporary rank binding file that augments global_env with
-    # any env_passthrough variables.
     with open(rank_binding_file) as f:
         rb = yaml.safe_load(f)
     rb.setdefault("global_env", {})
-    # Whitelist-based env passthrough (patterns) to avoid copying full env.
     default_env_patterns = ["VLLM_*", "MESH_DEVICE"]
     env_passthrough = override_tt_config.get("env_passthrough", default_env_patterns)
     if isinstance(env_passthrough, (list, tuple)):
@@ -217,9 +246,8 @@ def tt_run_launch(
                 if fnmatch.fnmatch(key, pattern):
                     to_inject[key] = val
                     break
-        # Do not override existing keys in global_env.
-        for k, v in to_inject.items():
-            rb["global_env"].setdefault(k, v)
+        for key, val in to_inject.items():
+            rb["global_env"].setdefault(key, val)
 
     tmp_rb_path = os.path.join(cfg_dir, "tmp_vllm_tt_rank_binding.yaml")
     with open(tmp_rb_path, "w") as tf:
@@ -232,8 +260,6 @@ def tt_run_launch(
                 "override_tt_config['extra_ttrun_args'] must be a string"
             )
         normalized_extra_ttrun_args = shlex.split(extra_ttrun_args)
-
-        # Prevent ambiguous/duplicated flags that vLLM is responsible for.
         reserved_flags = {"--rank-binding", "--mpi-args"}
         if any(
             (tok in reserved_flags)
@@ -242,22 +268,19 @@ def tt_run_launch(
         ):
             raise RuntimeError(
                 "override_tt_config['extra_ttrun_args'] must not include "
-                "--rank-binding or --mpi-args (vLLM sets these)"
+                "--rank-binding or --mpi-args"
             )
 
     cmd = ["tt-run"]
-    if normalized_extra_ttrun_args:
-        cmd.extend(normalized_extra_ttrun_args)
+    cmd.extend(normalized_extra_ttrun_args)
     cmd.extend(["--rank-binding", tmp_rb_path])
     if mpi_args:
-        # Pass raw string; tt-run will shlex.split it
         cmd.extend(["--mpi-args", mpi_args])
-    # Program to run per MPI rank: engine entrypoint with explicit args
     cmd.extend(
         [
             sys.executable,
             "-m",
-            "vllm.v1.engine.tt_core_launcher",
+            "vllm_tt_plugin.launcher",
             "--handshake",
             str(handshake_address),
             "--config-pkl",
@@ -267,32 +290,17 @@ def tt_run_launch(
         ]
     )
 
-    child_env = os.environ.copy()
-    pretty_cmd = shlex.join(cmd)
-    logger.info("Launching engines with tt-run: %s", pretty_cmd)
-    mpi_proc = subprocess.Popen(cmd, env=child_env)
-
-    # Set up finalizer for MPI subprocess cleanup
+    logger.info("Launching engines with tt-run: %s", shlex.join(cmd))
+    mpi_proc = subprocess.Popen(cmd, env=os.environ.copy())
     _setup_mpi_proc_finalizer(mpi_proc, cleanup_target)
 
 
 def _setup_mpi_proc_finalizer(
     mpi_proc: subprocess.Popen, cleanup_target: object
 ) -> None:
-    """
-    Set up a weakref finalizer to clean up MPI subprocess when cleanup_target
-    is garbage collected. This ensures graceful shutdown of TT-MPI processes.
-    Args:
-        mpi_proc: The subprocess.Popen object for the tt-run process
-        cleanup_target: Obj whose lifecycle decides when to clean up mpi_proc
-    """
-
     def _finalize_mpi(proc_ref):
         proc = proc_ref()
-        if proc is None:
-            return
-        # Check if process is already dead
-        if proc.poll() is not None:
+        if proc is None or proc.poll() is not None:
             return
         proc.terminate()
         try:
@@ -310,29 +318,16 @@ def _setup_mpi_proc_finalizer(
 def _parse_args() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="TT engine core entrypoint")
     parser.add_argument("--handshake", required=True, help="Handshake address")
-    parser.add_argument(
-        "--config-pkl",
-        required=True,
-        dest="config_pkl",
-        help="Path to serialized VllmConfig pickle",
-    )
-    parser.add_argument(
-        "--log-stats",
-        required=True,
-        choices=["0", "1"],
-        dest="log_stats",
-        help="Enable stat logging (1) or disable (0)",
-    )
+    parser.add_argument("--config-pkl", required=True, dest="config_pkl")
+    parser.add_argument("--log-stats", required=True, choices=["0", "1"])
     return parser.parse_args()
 
 
 def main() -> None:
     args = _parse_args()
-    handshake_address = args.handshake
-    config_pickle_path = args.config_pkl
-    log_stats = args.log_stats == "1"
+    with open(args.config_pkl, "rb") as f:
+        vllm_config: VllmConfig = cloudpickle.load(f)
 
-    # Derive MPI ranks if present (device ranks).
     has_mpi = "OMPI_COMM_WORLD_SIZE" in os.environ or "PMI_SIZE" in os.environ
     mpi_rank = int(
         os.environ.get("OMPI_COMM_WORLD_RANK", os.environ.get("PMI_RANK", "0"))
@@ -340,48 +335,25 @@ def main() -> None:
     mpi_world = int(
         os.environ.get("OMPI_COMM_WORLD_SIZE", os.environ.get("PMI_SIZE", "1"))
     )
-
-    # Load vllm config.
-    if not os.path.isfile(config_pickle_path):
-        raise RuntimeError(
-            f"Config file doesn't exist or isn't a file: {config_pickle_path}"
-        )
-    if os.path.islink(config_pickle_path):
-        raise RuntimeError(
-            f"Config file is a symlink (not allowed): {config_pickle_path}"
-        )
-    with open(config_pickle_path, "rb") as f:
-        vllm_config: VllmConfig = cloudpickle.load(f)
-
-    # Determine the global DP topology from the serialized config.
-    pc = vllm_config.parallel_config
-    dp_size = pc.data_parallel_size
-
-    # Map MPI ranks (subset) deterministically into global DP ranks:
-    # dp_rank = mpi_rank * (dp_size / mpi_world), assuming divisibility.
-    # This picks the first local DP rank on each host.
     if not has_mpi:
-        raise RuntimeError("tt_engine_core must be launched under MPI (tt-run)")
-    assert dp_size % mpi_world == 0, (
-        f"dp_size ({dp_size}) must be divisible by mpi_world ({mpi_world})"
-    )
-    segment = dp_size // mpi_world
-    pc.data_parallel_rank = mpi_rank * segment
-    pc.data_parallel_rank_local = 0  # Device processes use local DP rank 0.
+        raise RuntimeError("TT engine core must be launched under MPI")
 
-    # Ensure uniproc in engine process (worker inline).
+    pc = vllm_config.parallel_config
+    assert pc.data_parallel_size % mpi_world == 0
+    segment = pc.data_parallel_size // mpi_world
+    pc.data_parallel_rank = mpi_rank * segment
+    pc.data_parallel_rank_local = 0
     assert pc.distributed_executor_backend == "uni", (
         "TT MPI must be used with uniproc executor backend"
     )
 
-    # Run engine core busy loop.
-    # Local client is False since only non-device ranks are spawned in-process.
-    EngineCoreProc.run_engine_core(
+    engine_core_proc_cls = resolve_obj_by_qualname(pc.engine_core_proc_cls)
+    engine_core_proc_cls.run_engine_core(
         vllm_config=vllm_config,
         local_client=False,
-        handshake_address=handshake_address,
+        handshake_address=args.handshake,
         executor_class=UniProcExecutor,
-        log_stats=log_stats,
+        log_stats=args.log_stats == "1",
         dp_rank=pc.data_parallel_rank,
         local_dp_rank=pc.data_parallel_rank_local,
     )
@@ -393,5 +365,5 @@ if __name__ == "__main__":
     except SystemExit:
         raise
     except Exception:
-        logger.exception("tt_engine_core failed")
+        logger.exception("TT engine core failed")
         sys.exit(1)
